// src/lib/api/schema.d.ts
export interface components {
    schemas: {
        BBQReservation: {
            readonly id: number;
            reservation_user?: number | null;
            /** Format: date */
            reservation_date: string;
            guest_count?: number | null;
        };
        BBQReservationRequest: {
            reservation_user?: number | null;
            /** Format: date */
            reservation_date: string;
            guest_count?: number | null;
        };
        CondoPayment: {
            readonly id: number;
            title: string;
            status?: components["schemas"]["StatusEnum"];
            description?: string | null;
            payment_link: string;
            /** Format: decimal */
            amount?: string;
            /** Format: date */
            due_date?: string | null;
            /** Format: date-time */
            payment_date?: string | null;
            /** Format: date-time */
            readonly created_at: string;
            /** Format: date-time */
            readonly updated_at: string;
            payer_user: number;
        };
// ...
        HallReservation: {
            readonly id: number;
            reservation_user?: number | null;
            /** Format: date */
            reservation_date: string;
            guest_count?: number | null;
        };
// ...
        SunnyValeNews: {
            readonly id: number;
            title: string;
            description: string;
            author: string;
            priority_level?: components["schemas"]["PriorityLevelEnum"];
            /** Format: date-time */
            readonly created_at: string;
            /** Format: date-time */
            readonly updated_at: string;
        };
// ...
        User: {
            readonly id: number;
            /** @description Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only. */
            username: string;
            first_name: string;
            last_name: string;
            /**
             * Email address
             * Format: email
             */
            email: string;
        };
        UserRequest: {
            /** @description Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only. */
            username: string;
            password: string;
            first_name: string;
            last_name: string;
            /**
             * Email address
             * Format: email
             */
            email: string;
        };
        VisitorAccess: {
            readonly id: number;
            visitor_name: string;
            /** Format: email */
            email?: string | null;
            /** Format: date-time */
            scheduled_date: string;
            /** Format: date-time */
            readonly checkin_date_time: string | null;
            /** Format: date-time */
            checkout_date_time?: string | null;
            readonly checkin_code: string;
            readonly checkout_code: string;
            readonly status: string;
            /** Format: date-time */
            readonly created_at: string;
            /** Format: date-time */
            readonly updated_at: string;
            description?: string | null;
            readonly link_checkin: string | null;
            readonly link_checkout: string | null;
            host_user?: number | null;
        };
        VisitorAccessRequest: {
            visitor_name: string;
            /** Format: email */
            email?: string | null;
            /** Format: date-time */
            scheduled_date: string;
            /** Format: date-time */
            checkout_date_time?: string | null;
            description?: string | null;
            host_user?: number | null;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}